adding violin class to gui
